Create homebrew tap file

+++ scripts/release_homebrew.sh
@@ -1,38 +1,55 @@
 #!/bin/bash
+## Update the homebrew descriptor (Inspired by https://github.com/holgerbrandl/kscript/issues/50)
 
+FLOORPLAN_CLI_HOME=floorplan-cli
+FLOORPLAN_CLI_BINARY_NAME=floorplan-cli
 floorplan_version=$(grep 'floorPlanVersion' versioning.gradle | cut -f2 -d'=' | tr -d ' "')
+FLOORPLAN_DISTRIBUTION_NAME="${FLOORPLAN_CLI_BINARY_NAME}-${floorplan_version}"
 
 echo "Ready to release FloorPlan $floorplan_version to Homebrew."
 read -p "Do you wish to continue? " -n 1 -r
 echo    # move to a new line
-if [[ ! $REPLY =~ ^[Yy]$ ]]; thewn exit 1; fi
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then exit 1; fi
 
-## Update the homebrew descriptor (Inspired by https://github.com/holgerbrandl/kscript/issues/50)
+ARCHIVE_DIR=cli-archive/
+
+mkdir -p $ARCHIVE_DIR/
+cp ${FLOORPLAN_CLI_HOME}/build/distributions/${FLOORPLAN_DISTRIBUTION_NAME}.zip $ARCHIVE_DIR/
+
+if [ ! -f ${ARCHIVE_DIR}/${FLOORPLAN_DISTRIBUTION_NAME}.zip ]; then
+    echo "Distribution binary zip file not found!"
+    exit 1;
+fi
+
+archiveMd5=$(shasum -a 256 ${ARCHIVE_DIR}/${FLOORPLAN_DISTRIBUTION_NAME}.zip | cut -f1 -d ' ')
+
+echo "MD5 SHA256 of zipped binary is $archiveMd5."
+
 rm -rf homebrew-tap
 git clone https://github.com/julioz/homebrew-tap.git
 cd homebrew-tap
 
 git config user.email "julioz@users.noreply.github.com"
-#
-#archiveMd5=$(shasum -a 256 ${KSCRIPT_ARCHIVE}/kscript-${floorplan_version}.zip | cut -f1 -d ' ')
-#
-#cat - <<EOF > floorplan.rb
-#class FloorPlan < Formula
-#  desc "floorplan"
-#  homepage "https://github.com/julioz/FloorPlan"
-#  url "https://github.com/holgerbrandl/kscript/releases/download/v${floorplan_version}/kscript-${floorplan_version}-bin.zip"
-#  sha256 "${archiveMd5}"
-#
-#  depends_on "kotlin"
-#
-#  def install
-#    libexec.install Dir["*"] # Put the extracted JAR in to the 'private' libexec folder
-#    inreplace "#{libexec}/bin/kscript", /^jarPath=.*/, "jarPath=#{libexec}/bin/kscript.jar"
-#    bin.install_symlink "#{libexec}/bin/kscript"
-#  end
-#end
-#EOF
-#
+
+
+cat - <<EOF > floorplan.rb
+class FloorPlan < Formula
+  desc "floorplan"
+  homepage "https://github.com/julioz/FloorPlan"
+  url "https://github.com/holgerbrandl/kscript/releases/download/v${floorplan_version}/kscript-${floorplan_version}-bin.zip"
+  sha256 "${archiveMd5}"
+
+  bottle :unneeded
+
+  depends_on "kotlin"
+
+  def install
+    libexec.install Dir["*"] # Put the extracted files in to the 'private' libexec folder
+    bin.install_symlink "#{libexec}/bin/floorplan-cli"
+  end
+end
+EOF
+
 #git add kscript.rb
 #git commit -m "v${floorplan_version} release"
 #git push #origin releases
